Refactoring of permission classes #79
