+++ samples/Controllers/JwtBearerSample/Controllers/MeController.cs
@@ -1,6 +1,8 @@
 using System.Net.Mime;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using SimpleAuthentication.Permissions;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace JwtBearerSample.Controllers;
 
@@ -10,9 +12,11 @@ namespace JwtBearerSample.Controllers;
 public class MeController : ControllerBase
 {
     [Authorize]
+    [Permissions("profile")]
     [HttpGet]
     [ProducesResponseType(typeof(User), StatusCodes.Status200OK)]
     [ProducesDefaultResponseType]
+    [SwaggerOperation(description: "This endpoint requires the 'profile' permission")]
     public ActionResult<User> Get()
         => new User(User.Identity!.Name);
 }
