+++ samples/Controllers/JwtBearerSample/Controllers/PeopleController.cs
@@ -2,6 +2,7 @@ using System.Net.Mime;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using SimpleAuthentication.Permissions;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace JwtBearerSample.Controllers;
 
@@ -11,28 +12,43 @@ namespace JwtBearerSample.Controllers;
 [Produces(MediaTypeNames.Application.Json)]
 public class PeopleController : ControllerBase
 {
+    [Permissions(Permissions.PeopleRead, Permissions.PeopleAdmin)]
     [HttpGet]
+    [SwaggerOperation(description: $"This endpoint requires the '{Permissions.PeopleRead}' or '{Permissions.PeopleAdmin}' permissions")]
     public IActionResult GetList() => NoContent();
 
-    [Permission(Permissions.Read)]
+    [Permissions(Permissions.PeopleRead, Permissions.PeopleAdmin)]
     [HttpGet("{id:int}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesDefaultResponseType]
+    [SwaggerOperation(description: $"This endpoint requires the '{Permissions.PeopleRead}' or '{Permissions.PeopleAdmin}' permissions")]
     public IActionResult GetPerson(int id) => NoContent();
 
-    [Permission(Permissions.Write)]
+    [Permissions(Permissions.PeopleWrite)]
     [HttpPost]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesDefaultResponseType]
+    [SwaggerOperation(description: $"This endpoint requires the '{Permissions.PeopleWrite}' permission")]
     public IActionResult Insert() => NoContent();
 
+    [Permissions(Permissions.PeopleWrite)]
     [HttpPut]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesDefaultResponseType]
+    [SwaggerOperation(description: $"This endpoint requires the '{Permissions.PeopleWrite}' permission")]
     public IActionResult Update() => NoContent();
 
-    [Permission(Permissions.Write, Permissions.Admin)]
+    [Permissions(Permissions.PeopleAdmin)]
     [HttpDelete("{id:int}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesDefaultResponseType]
+    [SwaggerOperation(description: $"This endpoint requires the '{Permissions.PeopleAdmin}' permission")]
     public IActionResult Delete(int id) => NoContent();
 }
 
 public static class Permissions
 {
-    public const string Read = "read";
-    public const string Write = "write";
-    public const string Admin = "admin";
+    public const string PeopleRead = "people_read";
+    public const string PeopleWrite = "people_write";
+    public const string PeopleAdmin = "people_admin";
 }
