--- src/SimpleAuthentication.Abstractions/Permissions/IPermissionService.cs
@@ -1,8 +1,0 @@
-﻿using System.Security.Claims;
-
-namespace SimpleAuthentication.Permissions;
-
-public interface IPermissionService
-{
-    Task<bool> IsGrantedAsync(ClaimsPrincipal user, IEnumerable<string> permissions);
-}
