--- src/SimpleAuthentication.Abstractions/Permissions/PermissionAttribute.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-
-namespace SimpleAuthentication.Permissions;
-
-public class PermissionAttribute : AuthorizeAttribute
-{
-    public PermissionAttribute(params string[] permissions)
-        : base(string.Join(",", permissions))
-    {
-    }
-}
