+++ src/SimpleAuthentication/Permissions/PermissionAuthorizationHandler.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.AspNetCore.Authorization;
+
+namespace SimpleAuthentication.Permissions;
+
+internal class PermissionAuthorizationHandler : AuthorizationHandler<PermissionRequirement>
+{
+    private readonly IPermissionHandler permissionHandler;
+
+    public PermissionAuthorizationHandler(IPermissionHandler permissionHandler)
+    {
+        ArgumentNullException.ThrowIfNull(permissionHandler);
+
+        this.permissionHandler = permissionHandler;
+    }
+
+    protected override async Task HandleRequirementAsync(AuthorizationHandlerContext context, PermissionRequirement requirement)
+    {
+        var isGranted = await permissionHandler.IsGrantedAsync(context.User, requirement.Permissions);
+        if (isGranted)
+        {
+            context.Succeed(requirement);
+        }
+    }
+}
