+++ src/SimpleAuthentication/SimpleAuthenticationExtensions.cs
@@ -1,19 +1,14 @@
 ﻿using System.Text;
 using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.IdentityModel.Tokens;
-using Microsoft.Net.Http.Headers;
-using Microsoft.OpenApi.Models;
 using SimpleAuthentication.ApiKey;
 using SimpleAuthentication.BasicAuthentication;
 using SimpleAuthentication.JwtBearer;
-using SimpleAuthentication.Swagger;
-using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace SimpleAuthentication;
 
@@ -26,11 +21,11 @@ public static class SimpleAuthenticationExtensions
     /// <summary>
     /// Registers services required by authentication services, reading configuration from the specified <see cref="IConfiguration"/> source.
     /// </summary>
-    /// <param name="services">The <see cref="IServiceCollection"/>.</param>
+    /// <param name="services">The <see cref="IServiceCollection"/> to add services to.</param>
     /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
     /// <param name="sectionName">The name of the configuration section that holds authentication settings (default: Authentication).</param>
     /// <param name="addAuthorizationServices">Set to <see langword="true"/> to automatically add Authorization policy services.</param>
-    /// <returns>A <see cref="AuthenticationBuilder"/> that can be used to further customize authentication.</returns>
+    /// <returns>An <see cref="AuthenticationBuilder"/> that can be used to further customize authentication.</returns>
     /// <exception cref="ArgumentException">Configuration is invalid.</exception>
     /// <exception cref="ArgumentNullException">One or more required configuration settings are missing.</exception>
     /// <seealso cref="IServiceCollection"/>
@@ -56,11 +51,11 @@ public static class SimpleAuthenticationExtensions
     /// <summary>
     /// Registers services required by authentication services, reading configuration from the specified <see cref="IConfiguration"/> source.
     /// </summary>
-    /// <param name="builder">The <see cref="AuthenticationBuilder"/>.</param>
+    /// <param name="builder">The <see cref="AuthenticationBuilder"/> to add services to.</param>
     /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
     /// <param name="sectionName">The name of the configuration section that holds authentication settings (default: Authentication).</param>
     /// <param name="addAuthorizationServices">Set to <see langword="true"/> to automatically add Authorization policy services.</param>
-    /// <returns>A <see cref="AuthenticationBuilder"/> that can be used to further customize authentication.</returns>
+    /// <returns>An <see cref="AuthenticationBuilder"/> that can be used to further customize authentication.</returns>
     /// <exception cref="ArgumentException">Configuration is invalid.</exception>
     /// <exception cref="ArgumentNullException">One or more required configuration settings are missing.</exception>
     /// <seealso cref="AuthenticationBuilder"/>
@@ -141,7 +136,7 @@ public static class SimpleAuthenticationExtensions
 
             if (settings.ApiKeys.Any(k => string.IsNullOrWhiteSpace(k.Value) || string.IsNullOrWhiteSpace(k.UserName)))
             {
-                throw new ArgumentNullException("One or more API Keys contain null values");
+                throw new ArgumentNullException("Api Keys", "One or more API Keys contain null values");
             }
 
             builder.Services.Configure<ApiKeySettings>(section);
@@ -179,7 +174,7 @@ public static class SimpleAuthenticationExtensions
 
             if (settings.Credentials.Any(c => string.IsNullOrWhiteSpace(c.UserName) || string.IsNullOrWhiteSpace(c.Password)))
             {
-                throw new ArgumentNullException("One or more credentials contain null values");
+                throw new ArgumentNullException("Credentials", "One or more credentials contain null values");
             }
 
             builder.Services.Configure<BasicAuthenticationSettings>(section);
@@ -201,8 +196,8 @@ public static class SimpleAuthenticationExtensions
     /// must appear between the calls to <c>app.UseRouting()</c> and <c>app.UseEndpoints(...)</c> for
     /// the middleware to function correctly.
     /// </summary>
-    /// <param name="app">The <see cref="IApplicationBuilder"/> to add the middleware to.</param>
-    /// <returns>A reference to <paramref name="app"/> after the operation has completed.</returns>
+    /// <param name="app">The <see cref="IApplicationBuilder"/> to add middlewares to.</param>
+    /// <returns>A reference to this instance after the operation has completed.</returns>
     /// <seealso cref="AuthenticationMiddleware"/>
     /// <seealso cref="AuthorizationMiddleware"/>
     /// <seealso cref="IApplicationBuilder"/>
@@ -214,144 +209,5 @@ public static class SimpleAuthenticationExtensions
         app.UseAuthorization();
 
         return app;
-    }
-
-    /// <summary>
-    /// Adds authentication support in Swagger, enabling the Authorize button in the Swagger UI, reading configuration from the specified <see cref="IConfiguration"/> source.
-    /// </summary>
-    /// <param name="options">The <see cref="SwaggerGenOptions"/> to add configuration to.</param>
-    /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
-    /// <param name="sectionName">The name of the configuration section that holds authentication settings (default: Authentication).</param>
-    /// <seealso cref="SwaggerGenOptions"/>
-    /// <seealso cref="IConfiguration"/>
-    public static void AddSimpleAuthentication(this SwaggerGenOptions options, IConfiguration configuration, string sectionName = "Authentication")
-        => options.AddSimpleAuthentication(configuration, sectionName, Array.Empty<OpenApiSecurityRequirement>());
-
-    /// <summary>
-    /// Adds authentication support in Swagger, enabling the Authorize button in the Swagger UI, reading configuration from a section named <strong>Authentication</strong> in <see cref="IConfiguration"/> source.
-    /// </summary>
-    /// <param name="options">The <see cref="SwaggerGenOptions"/> to add configuration to.</param>
-    /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
-    /// <param name="additionalSecurityDefinitionNames">The name of additional security definitions that have been defined in Swagger.</param>
-    /// <seealso cref="SwaggerGenOptions"/>
-    /// <seealso cref="IConfiguration"/>
-    public static void AddSimpleAuthentication(this SwaggerGenOptions options, IConfiguration configuration, IEnumerable<string>? additionalSecurityDefinitionNames)
-        => options.AddSimpleAuthentication(configuration, "Authentication", additionalSecurityDefinitionNames);
-
-    /// <summary>
-    /// Adds authentication support in Swagger, enabling the Authorize button in the Swagger UI, reading configuration from the specified <see cref="IConfiguration"/> source.
-    /// </summary>
-    /// <param name="options">The <see cref="SwaggerGenOptions"/> to add configuration to.</param>
-    /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
-    /// <param name="sectionName">The name of the configuration section that holds authentication settings (default: Authentication).</param>
-    /// <param name="additionalSecurityDefinitionNames">The name of additional security definitions that have been defined in Swagger.</param>
-    /// <seealso cref="SwaggerGenOptions"/>
-    /// <seealso cref="IConfiguration"/>
-    public static void AddSimpleAuthentication(this SwaggerGenOptions options, IConfiguration configuration, string sectionName, IEnumerable<string>? additionalSecurityDefinitionNames)
-    {
-        var securityRequirements = additionalSecurityDefinitionNames?.Select(Helpers.CreateSecurityRequirement).ToArray();
-        options.AddSimpleAuthentication(configuration, sectionName, securityRequirements ?? Array.Empty<OpenApiSecurityRequirement>());
-    }
-
-    /// <summary>
-    /// Adds authentication support in Swagger, enabling the Authorize button in the Swagger UI, reading configuration from the specified <see cref="IConfiguration"/> source.
-    /// </summary>
-    /// <param name="options">The <see cref="SwaggerGenOptions"/> to add configuration to.</param>
-    /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
-    /// <param name="securityRequirements">Additional security requirements to be added to Swagger definition.</param>
-    /// <seealso cref="SwaggerGenOptions"/>
-    /// <seealso cref="IConfiguration"/>
-    public static void AddSimpleAuthentication(this SwaggerGenOptions options, IConfiguration configuration, params OpenApiSecurityRequirement[] securityRequirements)
-        => options.AddSimpleAuthentication(configuration, "Authentication", securityRequirements);
-
-    /// <summary>
-    /// Adds authentication support in Swagger, enabling the Authorize button in the Swagger UI, reading configuration from the specified <see cref="IConfiguration"/> source.
-    /// </summary>
-    /// <param name="options">The <see cref="SwaggerGenOptions"/> to add configuration to.</param>
-    /// <param name="configuration">The <see cref="IConfiguration"/> being bound.</param>
-    /// <param name="sectionName">The name of the configuration section that holds authentication settings (default: Authentication).</param>
-    /// <param name="additionalSecurityRequirements">Additional security requirements to be added to Swagger definition.</param>
-    /// <seealso cref="SwaggerGenOptions"/>
-    /// <seealso cref="IConfiguration"/>
-    public static void AddSimpleAuthentication(this SwaggerGenOptions options, IConfiguration configuration, string sectionName, params OpenApiSecurityRequirement[] additionalSecurityRequirements)
-    {
-        ArgumentNullException.ThrowIfNull(options);
-        ArgumentNullException.ThrowIfNull(configuration);
-        ArgumentNullException.ThrowIfNull(sectionName);
-
-        // Adds a security definition for each authentication method that has been configured.
-        CheckAddJwtBearer(options, configuration.GetSection($"{sectionName}:JwtBearer"));
-        CheckAddApiKey(options, configuration.GetSection($"{sectionName}:ApiKey"));
-        CheckAddBasicAuthentication(options, configuration.GetSection($"{sectionName}:Basic"));
-
-        if (additionalSecurityRequirements?.Any() ?? false)
-        {
-            // Adds all the other security requirements that have been specified.
-            foreach (var securityRequirement in additionalSecurityRequirements)
-            {
-                options.AddSecurityRequirement(securityRequirement);
-            }
-        }
-
-        options.OperationFilter<AuthenticationOperationFilter>();
-        options.DocumentFilter<ProblemDetailsDocumentFilter>();
-
-        static void CheckAddJwtBearer(SwaggerGenOptions options, IConfigurationSection section)
-        {
-            var settings = section.Get<JwtBearerSettings>();
-            if (settings is null)
-            {
-                return;
-            }
-
-            AddSecurityDefinition(options, settings.SchemeName, SecuritySchemeType.Http, JwtBearerDefaults.AuthenticationScheme, ParameterLocation.Header, HeaderNames.Authorization, "Insert the Bearer Token");
-            AddSecurityRequirement(options, settings.SchemeName);
-        }
-
-        static void CheckAddApiKey(SwaggerGenOptions options, IConfigurationSection section)
-        {
-            var settings = section.Get<ApiKeySettings>();
-            if (settings is null)
-            {
-                return;
-            }
-
-            if (!string.IsNullOrWhiteSpace(settings.HeaderName))
-            {
-                AddSecurityDefinition(options, $"{settings.SchemeName} in Header", SecuritySchemeType.ApiKey, null, ParameterLocation.Header, settings.HeaderName, "Insert the API Key");
-                AddSecurityRequirement(options, $"{settings.SchemeName} in Header");
-            }
-
-            if (!string.IsNullOrWhiteSpace(settings.QueryStringKey))
-            {
-                AddSecurityDefinition(options, $"{settings.SchemeName} in Query String", SecuritySchemeType.ApiKey, null, ParameterLocation.Query, settings.QueryStringKey, "Insert the API Key");
-                AddSecurityRequirement(options, $"{settings.SchemeName} in Query String");
-            }
-        }
-
-        static void CheckAddBasicAuthentication(SwaggerGenOptions options, IConfigurationSection section)
-        {
-            var settings = section.Get<BasicAuthenticationSettings>();
-            if (settings is null)
-            {
-                return;
-            }
-
-            AddSecurityDefinition(options, settings.SchemeName, SecuritySchemeType.Http, BasicAuthenticationDefaults.AuthenticationScheme, ParameterLocation.Header, HeaderNames.Authorization, "Insert user name and password");
-            AddSecurityRequirement(options, settings.SchemeName);
-        }
-
-        static void AddSecurityDefinition(SwaggerGenOptions options, string name, SecuritySchemeType securitySchemeType, string? scheme, ParameterLocation location, string parameterName, string description)
-            => options.AddSecurityDefinition(name, new()
-            {
-                In = location,
-                Name = parameterName,
-                Description = description,
-                Type = securitySchemeType,
-                Scheme = scheme
-            });
-
-        static void AddSecurityRequirement(SwaggerGenOptions options, string name)
-            => options.AddSecurityRequirement(Helpers.CreateSecurityRequirement(name));
     }
 }
